+++ src/data/index.ts
@@ -7,5 +7,6 @@ export * from './colors';
 export * from './contact';
 export * from './interests';
 export * from './services';
+export * from './shareButtons';
 export * from './skills';
 export * from './timeline';